pyproject.toml
[project]
name = "collab-book-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "alembic-postgresql-enum>=1.8.0",
    "asyncpg>=0.31.0",
    "edwh-uuid7>=0.2.3",
    "fastapi[standard]>=0.128.0",
    "fonttools>=4.61.1",
    "pillow>=12.1.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.6",
    "pymupdf-layout>=1.26.6",
    "pymupdf4llm>=0.2.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "tenacity>=9.1.4",
]
